pyproject.toml
[tool.poetry]
name = "ai-email-assistant"
version = "0.1.0"
description = "AI Email Assistant з Gmail Integration"
authors = ["Admin <PyDevDeep>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.135.3"
uvicorn = "^0.29.0"
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.30.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
structlog = "^24.1.0"
tenacity = "^8.2.3"
google-api-python-client = "^2.124.0"
google-auth-oauthlib = "^1.2.0"
openai = "^1.16.2"
anthropic = "^0.23.1"
slowapi = "^0.1.9"
prometheus-fastapi-instrumentator = "^7.0.0"
orjson = "^3.10.0"
apscheduler = "^3.10.4"
prometheus-client = "^0.25.0"
celery-prometheus-exporter = "^1.7.0"
beautifulsoup4 = "^4.14.3"
python-multipart = ">=0.0.26"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
fakeredis = "^2.22.0"
ruff = ">=0.9.5"
pyright = ">=1.1.400"
pre-commit = "^3.7.0"
safety = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore:Support for class-based `config` is deprecated:DeprecationWarning"
]
markers = [
    "unit: fast isolated unit tests",
    "integration: tests that touch real external services",
]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    "venv", ".venv", "__pycache__", "build", "dist",
    "logs", "alembic", ".pytest_cache", "htmlcov", ".coverage"
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "S", "W", "N", "UP", "C4", "SIM", "RUF", "A", "DTZ", "TID", "ASYNC", "PT"]
ignore = ["F401", "E501", "B008","RUF001", "RUF002", "RUF003", "N815", "S107"]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "N806", "SIM117", "PT011", "PT006"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/.*", "**/__pycache__", "build", "dist", "venv", ".venv", "alembic"]
pythonVersion = "3.11"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownMemberType = false